CBTUI-64 - remove unused methods, pages, components
