--- src/model/response/usercrypto/PageUserCryptoResponse.ts
@@ -1,8 +1,0 @@
-import {UserCryptoResponse} from "./UserCryptoResponse";
-
-export interface PageUserCryptoResponse {
-  page: number,
-  totalPages: number,
-  hasNextPage: boolean,
-  cryptos: Array<UserCryptoResponse>
-}
