+++ src/pages/crypto/CryptosPage.tsx
@@ -1,214 +1,158 @@
 import Navbar from "../../components/page/Navbar";
 import Footer from "../../components/page/Footer";
-import AddNewButton from "../../components/buttons/AddNewButton";
+import React, {Fragment} from "react";
 import withScrollToTop from "../../hoc/withScrollToTop";
-import React, {useEffect, useRef, useState} from "react";
-import {UserCryptoResponse} from "../../model/response/usercrypto/UserCryptoResponse";
-import {PageUserCryptoResponse} from "../../model/response/usercrypto/PageUserCryptoResponse";
-import {deleteCryptoService, getCryptosByPageService} from "../../services/cryptoService";
-import TableSkeleton from "../../components/skeletons/TableSkeleton";
-import ErrorAlert from "../../components/page/ErrorAlert";
-import {SortableTableColumnTitle} from "../../components/table/SortableTableColumnTitle";
-import {TableColumnTitle} from "../../components/table/TableColumnTitle";
-import {TableColumnContent} from "../../components/table/TableColumnContent";
-import EditButton from "../../components/table/EditButton";
-import TransferButton from "../../components/table/TransferButton";
-import DeleteButton from "../../components/table/DeleteButton";
+import {useSortParams} from "../../hooks/useSortParams";
+import {usePageUserCryptosInsightsResponse} from "../../hooks/usePageUserCryptosInsightsResponse";
+import {retrieveCryptosPlatformsInsightsByPage} from "../../services/insightsService";
+import {PageUserCryptosInsightsResponse} from "../../model/response/insight/PageUserCryptosInsightsResponse";
+import NoCryptosFoundAlert from "../../components/crypto/NoCryptosFoundAlert";
 import LoadMoreButton from "../../components/buttons/LoadMoreButton";
-import FilterField from "../../components/commons/FilterField";
+import InsightsSortFilterComponent from "../../components/insights/InsightsSortFilterComponent";
+import {Link} from "react-router-dom";
+import TableSkeleton from "../../components/skeletons/TableSkeleton";
 
 const CryptosPage = () => {
 
-  const filteredCryptos = useRef<Array<UserCryptoResponse>>([]);
-  const [pageUserCryptoResponse, setPageUserCryptoResponse] = useState<PageUserCryptoResponse>({
-    page: 0,
-    totalPages: 0,
-    hasNextPage: false,
-    cryptos: []
-  });
-  const [filterValue, setFilterValue] = useState("");
-  const [isLoadingMoreCryptos, setIsLoadingMoreCryptos] = useState(false);
-  const [page, setPage] = useState(0);
-  const [error, setError] = useState(false);
-  const [isLoadingCryptos, setIsLoadingCryptos] = useState(true);
-  const [sortAscending, setSortAscending] = useState(true);
+  const {sortParams, updateSortBy, updateSortType} = useSortParams();
+  const {
+    pageUserCryptosInsightsResponse,
+    setPageUserCryptosInsightsResponse,
+    filteredCryptos,
+    page,
+    setPage,
+    error,
+    isLoadingUserCryptosInsights,
+    isLoadingMore,
+    loadMoreCryptos,
+    filterTable,
+    cryptosFilterValue,
+    setCryptosFilterValue
+  } = usePageUserCryptosInsightsResponse(() => retrieveCryptosPlatformsInsightsByPage(0, sortParams));
 
-  useEffect(() => {
-    (async () => {
-        try {
-          const response = await getCryptosByPageService(page);
+  const retrieveSortedResults = async () => {
+    setCryptosFilterValue("");
+    const response: PageUserCryptosInsightsResponse = await retrieveCryptosPlatformsInsightsByPage(0, sortParams);
+    setPage(0);
+    setPageUserCryptosInsightsResponse(response);
+    filteredCryptos.current = response.cryptos;
+  }
 
-          setPageUserCryptoResponse(response);
-          filteredCryptos.current = response.cryptos;
-        } catch (error: unknown) {
-          setError(true);
-        } finally {
-          setIsLoadingCryptos(false);
-        }
-      }
-    )();
-  }, []);
-
-  const deleteCrypto = async (cryptoId: string) => {
-    try {
-      await deleteCryptoService(cryptoId);
-
-      const updatedFilteredCryptos = filteredCryptos.current.filter(crypto => crypto.id !== cryptoId);
-      const updatedCryptos = pageUserCryptoResponse.cryptos.filter(crypto => crypto.id !== cryptoId);
-      const {hasNextPage, page, totalPages} = pageUserCryptoResponse;
-
-      setPageUserCryptoResponse({
-        cryptos: updatedCryptos,
-        hasNextPage,
-        page,
-        totalPages
-      });
-      filteredCryptos.current = updatedFilteredCryptos;
-    } catch (error: unknown) {
-      setError(true);
+  const returnChangePercentageColor = (changePercentage: string) => {
+    if (Number(changePercentage) < 0) {
+      return "text-red-500";
     }
-  }
 
-  const filterTable = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value;
-    filteredCryptos.current = pageUserCryptoResponse.cryptos.filter(crypto => doesMatchFilter(crypto, value));
-    setFilterValue(value);
-  }
-
-  const loadMoreCryptos = async () => {
-    setIsLoadingMoreCryptos(true);
-    const nextPage = page + 1;
-    setPage(nextPage);
-
-    try {
-      const response = await getCryptosByPageService(nextPage);
-      filteredCryptos.current = [...filteredCryptos.current, ...response.cryptos.filter((crypto) => doesMatchFilter(crypto, filterValue))];
-
-      setPageUserCryptoResponse({
-        cryptos: [...pageUserCryptoResponse.cryptos, ...response.cryptos],
-        hasNextPage: response.hasNextPage,
-        page: response.page,
-        totalPages: response.totalPages
-      });
-    } catch (error: unknown) {
-      setError(true);
-    } finally {
-      setIsLoadingMoreCryptos(false);
+    if (Number(changePercentage) > 0) {
+      return "text-green-500"
     }
-  }
 
-  const doesMatchFilter = (userCrypto: UserCryptoResponse, value: string) => {
-    return userCrypto.cryptoName.toUpperCase().startsWith(value.toUpperCase()) ||
-      userCrypto.platform.toUpperCase().startsWith(value.toUpperCase());
-  }
-
-  const sortByCryptoName = () => {
-    filteredCryptos.current = filteredCryptos.current.toSorted((a, b) => {
-      if (a.cryptoName > b.cryptoName) {
-        return sortAscending ? 1 : -1;
-      }
-
-      if (b.cryptoName > a.cryptoName) {
-        return sortAscending ? -1 : 1;
-      }
-
-      return 0;
-    });
-    setSortAscending(!sortAscending);
-  }
-
-  const sortByPlatformName = () => {
-    filteredCryptos.current = filteredCryptos.current.toSorted((a, b) => {
-      if (a.platform > b.platform) {
-        return sortAscending ? 1 : -1;
-      }
-
-      if (b.platform > a.platform) {
-        return sortAscending ? -1 : 1;
-      }
-
-      return 0;
-    });
-    setSortAscending(!sortAscending);
+    return "text-gray-50"
   }
 
   return (
-    <>
+    <Fragment>
       <Navbar/>
-      <div className="flex flex-col items-center min-h-screen">
-        <AddNewButton text="+ Add New Crypto" href="/crypto"/>
-        {
-          isLoadingCryptos && !error &&
-          <TableSkeleton/>
-        }
+      {
+        (pageUserCryptosInsightsResponse.cryptos === undefined || pageUserCryptosInsightsResponse.cryptos?.length === 0)
+        && !isLoadingUserCryptosInsights &&
+        <div className="min-h-screen">
+          <NoCryptosFoundAlert/>
+        </div>
+      }
 
-        {
-          error && !isLoadingCryptos &&
-          <ErrorAlert message="Error retrieving cryptos"/>
-        }
+      {
+        isLoadingUserCryptosInsights && !error &&
+        <TableSkeleton/>
+      }
 
-        {
-          !error && !isLoadingCryptos && pageUserCryptoResponse?.cryptos?.length > 0 &&
-          <div className="relative overflow-x-auto shadow-md sm:rounded-lg m-10 w-11/12">
-            <FilterField filterFunction={event => filterTable(event)}
-                         filterValue={filterValue}
-                         placeHolder="Search by crypto or platform"/>
+      {
+        !error && !isLoadingUserCryptosInsights && pageUserCryptosInsightsResponse.cryptos?.length > 0 &&
+        <div className="mx-10 max-w-[1920px] min-h-screen">
+          <InsightsSortFilterComponent filterFunction={filterTable}
+                                       filterValue={cryptosFilterValue}
+                                       updateSortBy={updateSortBy}
+                                       updateSortType={updateSortType}
+                                       retrieveSortedResults={retrieveSortedResults}/>
 
-            <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-              <thead className="text-xs text-gray-900 uppercase bg-gray-200 dark:bg-gray-700 dark:text-gray-400">
+          <div className="relative overflow-x-auto rounded-lg mt-10">
+            <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+              <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
               <tr>
-                <SortableTableColumnTitle title="Crypto"
-                                          additionalClasses="text-center"
-                                          sortFunction={sortByCryptoName}/>
-                <TableColumnTitle title="Quantity"
-                                  additionalClasses="text-center"/>
-                <SortableTableColumnTitle title="Platform"
-                                          additionalClasses="text-center"
-                                          sortFunction={sortByPlatformName}/>
-                <TableColumnTitle title="Action"
-                                  additionalClasses="text-center"/>
+                <th scope="col" className="px-6 py-3">#</th>
+                <th scope="col" className="px-6 py-3">Crypto</th>
+                <th scope="col" className="px-6 py-3">Percentage</th>
+                <th scope="col" className="px-6 py-3">Balance</th>
+                <th scope="col" className="px-6 py-3">Quantity</th>
+                <th scope="col" className="px-6 py-3">Price</th>
+                <th scope="col" className="px-6 py-3">24H Change</th>
+                <th scope="col" className="px-6 py-3">7D Change</th>
+                <th scope="col" className="px-6 py-3">30D Change</th>
+                <th scope="col" className="px-6 py-3">View</th>
               </tr>
               </thead>
               <tbody>
-              {
-                filteredCryptos.current.map(crypto => {
-                  const {id, cryptoName, platform, quantity} = crypto;
-
-                  return (
-                    <tr key={id}
-                        className="bg-gray-100 border-b dark:bg-gray-900 hover:bg-gray-50 dark:hover:bg-gray-800 dark:border-gray-700">
-                      <TableColumnContent content={cryptoName}
-                                          rowScope={true}
-                                          additionalClasses="text-center"/>
-                      <TableColumnContent content={quantity.toString()}
-                                          additionalClasses="text-center"/>
-                      <TableColumnContent content={platform}
-                                          additionalClasses="text-center"/>
-                      <td className="px-6 py-4 text-center flex flex-col justify-center space-y-2 lg:space-y-0 lg:space-x-4 lg:flex-row">
-                        <EditButton editLink={`/crypto/${id}`}/>
-                        <TransferButton transferLink={`/transfer/${id}`}/>
-                        <DeleteButton deleteFunction={() => deleteCrypto(id)}
-                                      deleteId={id}
-                                      deleteMessage={`Are you sure you want to delete ${cryptoName.toUpperCase()} in ${platform}?`}/>
-                      </td>
-                    </tr>
-                  );
-                })
-              }
+              {filteredCryptos.current.map((crypto, index) => (
+                <tr
+                  className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
+                  <th scope="row" className="px-6 py-4">{index + 1}</th>
+                  <td className="flex items-center px-6 py-4 text-gray-900 whitespace-nowrap dark:text-white">
+                    <img className="w-10 h-10 rounded-full" src={crypto.cryptoInfo.image}
+                         alt={`${crypto.cryptoInfo.cryptoName} logo`}/>
+                    <div className="pl-3">
+                      <div className="text-base font-semibold">
+                        {crypto.cryptoInfo.symbol.toUpperCase()}
+                      </div>
+                      <div
+                        className="font-normal text-gray-500 max-w-[120px] overflow-hidden text-ellipsis whitespace-nowrap">
+                        {crypto.cryptoInfo.cryptoName}
+                      </div>
+                    </div>
+                  </td>
+                  <td className="px-6 py-4">{`${crypto.percentage}%`}</td>
+                  <td className="px-6 py-4">{`$${crypto.balances.totalUSDBalance}`}</td>
+                  <td className="px-6 py-4">{crypto.quantity}</td>
+                  <td className="px-6 py-4">{`$${crypto.marketData.currentPrice.usd}`}</td>
+                  <td
+                    className={`px-6 py-4 ${returnChangePercentageColor(crypto.marketData.priceChange.changePercentageIn24h)}`}>
+                    {`${crypto.marketData.priceChange.changePercentageIn24h}%`}
+                  </td>
+                  <td
+                    className={`px-6 py-4 ${returnChangePercentageColor(crypto.marketData.priceChange.changePercentageIn7d)}`}>
+                    {`${crypto.marketData.priceChange.changePercentageIn7d}%`}
+                  </td>
+                  <td
+                    className={`px-6 py-4 ${returnChangePercentageColor(crypto.marketData.priceChange.changePercentageIn30d)}`}>
+                    {`${crypto.marketData.priceChange.changePercentageIn30d}%`}
+                  </td>
+                  <td className="px-6 py-4">
+                    <Link to={`/insights/cryptos/${crypto.cryptoInfo.cryptoId}`}>
+                      <button
+                        type="button"
+                        className="flex items-center justify-center w-full text-green-700 hover:text-white border border-green-700 hover:bg-green-800 focus:ring-4 focus:outline-none focus:ring-green-300 font-medium rounded-lg text-sm px-4 py-2 dark:border-green-500 dark:text-green-500 dark:hover:text-white dark:hover:bg-green-500 dark:focus:ring-green-800"
+                      >
+                        View
+                      </button>
+                    </Link>
+                  </td>
+                </tr>
+              ))}
               </tbody>
             </table>
-          </div>
-        }
 
-        {
-          !error && !isLoadingCryptos && pageUserCryptoResponse.hasNextPage &&
-          <LoadMoreButton loadMoreCallback={() => loadMoreCryptos()}
-                          isLoadingMore={isLoadingMoreCryptos}/>
-        }
-      </div>
+          </div>
+        </div>
+      }
+
+      {
+        !error && !isLoadingUserCryptosInsights && pageUserCryptosInsightsResponse.hasNextPage &&
+        <LoadMoreButton
+          loadMoreCallback={() => loadMoreCryptos(retrieveCryptosPlatformsInsightsByPage(page + 1, sortParams))}
+          isLoadingMore={isLoadingMore}/>
+      }
       <Footer/>
-    </>
-  );
+    </Fragment>
+  )
 }
 
 export default withScrollToTop(CryptosPage)
